fix: break the scan/duplicates/report import cycle
scan.py imported duplicates and report at the bottom while both imported
scan at the top, so `import croissant_baker.report` raised ImportError
unless something had already imported scan. Both modules are documented
import paths in docs/reference/api.md.

Move the shared vocabulary — Outcome, Reason, REASON_LABELS,
UNRESOLVED_OUTCOMES, ScanEntry — into a new leaf module, entries.py, that
imports nothing from the package. duplicates and report import the leaf;
scan keeps scan_directory and re-exports every name it exported before, so
downstream import paths are unchanged.

An in-process test cannot catch this, since sys.modules is already
populated by the time the suite runs, so the guard is a subprocess import
per module.
